OPS-3094 Chenages dur to review comment
